[frontend] refacto
