+++ opencti-platform/opencti-front/src/private/components/common/containers/ContainerHeader.jsx
@@ -4,24 +4,11 @@ import Typography from '@mui/material/Typography';
 import { Link } from 'react-router-dom';
 import Tooltip from '@mui/material/Tooltip';
 import { ChartTimeline, VectorLink, VectorPolygon } from 'mdi-material-ui';
-import { AddTaskOutlined, AssistantOutlined, ViewColumnOutlined } from '@mui/icons-material';
+import { ViewColumnOutlined } from '@mui/icons-material';
 import ToggleButton from '@mui/material/ToggleButton';
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
-import { DialogTitle } from '@mui/material';
-import Dialog from '@mui/material/Dialog';
-import DialogContent from '@mui/material/DialogContent';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import Badge from '@mui/material/Badge';
-import ListItemText from '@mui/material/ListItemText';
-import ListItemSecondaryAction from '@mui/material/ListItemSecondaryAction';
-import IconButton from '@mui/material/IconButton';
-import DialogActions from '@mui/material/DialogActions';
-import Button from '@mui/material/Button';
-import Select from '@mui/material/Select';
-import MenuItem from '@mui/material/MenuItem';
-import CircularProgress from '@mui/material/CircularProgress';
 import { makeStyles, useTheme } from '@mui/styles';
+import StixCoreObjectsSuggestions from '../stix_core_objects/StixCoreObjectsSuggestions';
 import { DraftChip } from '../draft/DraftChip';
 import { stixCoreObjectQuickSubscriptionContentQuery } from '../stix_core_objects/stixCoreObjectTriggersUtils';
 import StixCoreObjectAskAI from '../stix_core_objects/StixCoreObjectAskAI';
@@ -33,21 +20,11 @@ import useHelper from '../../../../utils/hooks/useHelper';
 import Security from '../../../../utils/Security';
 import { useFormatter } from '../../../../components/i18n';
 import { truncate } from '../../../../utils/String';
-import { commitMutation, MESSAGING$, QueryRenderer } from '../../../../relay/environment';
 import { getMainRepresentative } from '../../../../utils/defaultRepresentatives';
-import { stixCoreRelationshipCreationMutation } from '../stix_core_relationships/StixCoreRelationshipCreation';
-import { containerAddStixCoreObjectsLinesRelationAddMutation } from './ContainerAddStixCoreObjectsLines';
 import StixCoreObjectSharing from '../stix_core_objects/StixCoreObjectSharing';
-import useGranted, {
-  KNOWLEDGE_KNENRICHMENT,
-  KNOWLEDGE_KNGETEXPORT_KNASKEXPORT,
-  KNOWLEDGE_KNUPDATE,
-  KNOWLEDGE_KNUPDATE_KNMANAGEAUTHMEMBERS,
-} from '../../../../utils/hooks/useGranted';
+import { KNOWLEDGE_KNENRICHMENT, KNOWLEDGE_KNGETEXPORT_KNASKEXPORT, KNOWLEDGE_KNUPDATE, KNOWLEDGE_KNUPDATE_KNMANAGEAUTHMEMBERS } from '../../../../utils/hooks/useGranted';
 import StixCoreObjectQuickSubscription from '../stix_core_objects/StixCoreObjectQuickSubscription';
-import MarkdownDisplay from '../../../../components/MarkdownDisplay';
 import StixCoreObjectFileExport from '../stix_core_objects/StixCoreObjectFileExport';
-import Transition from '../../../../components/Transition';
 import { authorizedMembersToOptions, useGetCurrentUserAccessRight } from '../../../../utils/authorizedMembers';
 import StixCoreObjectEnrichment from '../stix_core_objects/StixCoreObjectEnrichment';
 
@@ -471,265 +448,15 @@ const ContainerHeader = (props) => {
   const theme = useTheme();
   const { t_i18n, fd } = useFormatter();
   const { isFeatureEnable } = useHelper();
-  const userIsKnowledgeEditor = useGranted([KNOWLEDGE_KNUPDATE]);
-  const [displaySuggestions, setDisplaySuggestions] = useState(false);
-  const [selectedEntity, setSelectedEntity] = useState({});
-  const [applying, setApplying] = useState([]);
-  const [applied, setApplied] = useState([]);
+
   const [displayEnrichment, setDisplayEnrichment] = useState(false);
   const isFABReplaced = isFeatureEnable('FAB_REPLACEMENT');
-  // Suggestions
-  const resolveThreats = (objects) => objects.filter(
-    (o) => [
-      'Threat-Actor',
-      'Intrusion-Set',
-      'Campaign',
-      'Incident',
-      'Malware',
-      'Tool',
-    ].includes(o.entity_type) && o.types.includes('manual'),
-  );
-  const resolveIndicators = (objects) => objects.filter(
-    (o) => ['Indicator'].includes(o.entity_type) && o.types.includes('manual'),
-  );
-  const resolveArsenal = (objects) => objects.filter(
-    (o) => ['Attack-Pattern', 'Malware', 'Tool', 'Channel', 'Narrative'].includes(
-      o.entity_type,
-    ) && o.types.includes('manual'),
-  );
-  const resolveTargets = (objects) => objects.filter(
-    (o) => [
-      'Sector',
-      'Region',
-      'Country',
-      'City',
-      'Position',
-      'Organization',
-      'System',
-      'Individual',
-      'Vulnerability',
-    ].includes(o.entity_type) && o.types.includes('manual'),
-  );
-  const setAppliedSuggestion = (suggestion) => {
-    const appliedSuggestions = JSON.parse(
-      localStorage.getItem(`suggestions-rules-${container.id}`) || '[]',
-    );
-    localStorage.setItem(
-      `suggestions-rules-${container.id}`,
-      JSON.stringify([...appliedSuggestions, suggestion]),
-    );
-  };
-  const getAppliedSuggestions = () => {
-    return JSON.parse(
-      localStorage.getItem(`suggestions-rules-${container.id}`) || '[]',
-    );
-  };
+
   const handleCloseEnrichment = () => {
     setDisplayEnrichment(false);
   };
   const handleOpenEnrichment = () => {
     setDisplayEnrichment(true);
-  };
-  const generateSuggestions = (objects) => {
-    const suggestions = [];
-    const resolvedThreats = resolveThreats(objects);
-    const resolvedIndicators = resolveIndicators(objects);
-    const resolvedArsenal = resolveArsenal(objects);
-    const resolvedTargets = resolveTargets(objects);
-    // Threats and indicators
-    if (resolvedThreats.length > 0 && resolvedIndicators.length > 0) {
-      suggestions.push({ type: 'threats-indicators', data: resolvedThreats });
-    }
-    // Threats and arsenal
-    if (resolvedThreats.length > 0 && resolvedArsenal.length > 0) {
-      suggestions.push({ type: 'threats-arsenal', data: resolvedThreats });
-    }
-    // Threats and targets
-    if (resolvedThreats.length > 0 && resolvedTargets.length > 0) {
-      suggestions.push({ type: 'threats-targets', data: resolvedThreats });
-    }
-    return suggestions;
-  };
-  const handleSelectEntity = (type, event) => {
-    if (event && event.target && event.target.value) {
-      setSelectedEntity({ ...selectedEntity, [type]: event.target.value });
-    }
-  };
-  // eslint-disable-next-line consistent-return
-  const applySuggestion = async (type, objects) => {
-    if (type === 'threats-indicators' && selectedEntity) {
-      // create all indicates relationships
-      setApplying([...applying, type]);
-      const resolvedIndicators = resolveIndicators(objects);
-      const createdRelationships = await Promise.all(
-        resolvedIndicators.map((indicator) => {
-          const values = {
-            relationship_type: 'indicates',
-            confidence: container.confidence,
-            fromId: indicator.id,
-            toId: selectedEntity[type],
-            createdBy: container.createdBy?.id,
-            objectMarking: container.objectMarking.map((m) => m.id),
-          };
-          return new Promise((resolve) => {
-            commitMutation({
-              mutation: stixCoreRelationshipCreationMutation,
-              variables: {
-                input: values,
-              },
-              onCompleted: (response) => resolve(response.stixCoreRelationshipAdd),
-            });
-          });
-        }),
-      );
-      await Promise.all(
-        createdRelationships.map((createdRelationship) => {
-          const input = {
-            toId: createdRelationship.id,
-            relationship_type: 'object',
-          };
-          return new Promise((resolve) => {
-            commitMutation({
-              mutation: containerAddStixCoreObjectsLinesRelationAddMutation,
-              variables: {
-                id: container.id,
-                input,
-              },
-              onCompleted: (response) => resolve(response.containerEdit.relationAdd),
-            });
-          });
-        }),
-      );
-      MESSAGING$.notifySuccess('Suggestion successfully applied.');
-      setAppliedSuggestion(type);
-      setApplied([
-        ...applied,
-        {
-          [type]: selectedEntity[type],
-        },
-      ]);
-      setApplying(applying.filter((n) => n !== type));
-      if (onApplied) {
-        return onApplied(createdRelationships);
-      }
-    }
-    if (type === 'threats-arsenal' && selectedEntity) {
-      // create all targets relationships
-      setApplying([...applying, type]);
-      const selectedObjectType = objects
-        .filter((n) => n.id === selectedEntity[type])
-        .at(0).entity_type;
-      const resolvedArsenal = resolveArsenal(objects).filter(
-        (n) => n.entity_type !== selectedObjectType,
-      );
-      const createdRelationships = await Promise.all(
-        resolvedArsenal.map((arsenal) => {
-          const values = {
-            relationship_type: 'uses',
-            confidence: container.confidence,
-            fromId: selectedEntity[type],
-            toId: arsenal.id,
-            createdBy: container.createdBy?.id,
-            objectMarking: container.objectMarking.map((m) => m.id),
-          };
-          return new Promise((resolve) => {
-            commitMutation({
-              mutation: stixCoreRelationshipCreationMutation,
-              variables: {
-                input: values,
-              },
-              onCompleted: (response) => resolve(response.stixCoreRelationshipAdd),
-            });
-          });
-        }),
-      );
-      await Promise.all(
-        createdRelationships.map((createdRelationship) => {
-          const input = {
-            toId: createdRelationship.id,
-            relationship_type: 'object',
-          };
-          return new Promise((resolve) => {
-            commitMutation({
-              mutation: containerAddStixCoreObjectsLinesRelationAddMutation,
-              variables: {
-                id: container.id,
-                input,
-              },
-              onCompleted: (response) => resolve(response.containerEdit.relationAdd),
-            });
-          });
-        }),
-      );
-      MESSAGING$.notifySuccess('Suggestion successfully applied.');
-      setAppliedSuggestion(type);
-      setApplied([
-        ...applied,
-        {
-          [type]: selectedEntity[type],
-        },
-      ]);
-      setApplying(applying.filter((n) => n !== type));
-      if (onApplied) {
-        return onApplied(createdRelationships);
-      }
-    }
-    if (type === 'threats-targets' && selectedEntity) {
-      // create all targets relationships
-      setApplying([...applying, type]);
-      const resolvedTargets = resolveTargets(objects);
-      const createdRelationships = await Promise.all(
-        resolvedTargets.map((target) => {
-          const values = {
-            relationship_type: 'targets',
-            confidence: container.confidence,
-            fromId: selectedEntity[type],
-            toId: target.id,
-            createdBy: container.createdBy?.id,
-            objectMarking: container.objectMarking.map((m) => m.id),
-          };
-          return new Promise((resolve) => {
-            commitMutation({
-              mutation: stixCoreRelationshipCreationMutation,
-              variables: {
-                input: values,
-              },
-              onCompleted: (response) => resolve(response.stixCoreRelationshipAdd),
-            });
-          });
-        }),
-      );
-      await Promise.all(
-        createdRelationships.map((createdRelationship) => {
-          const input = {
-            toId: createdRelationship.id,
-            relationship_type: 'object',
-          };
-          return new Promise((resolve) => {
-            commitMutation({
-              mutation: containerAddStixCoreObjectsLinesRelationAddMutation,
-              variables: {
-                id: container.id,
-                input,
-              },
-              onCompleted: (response) => resolve(response.containerEdit.relationAdd),
-            });
-          });
-        }),
-      );
-      MESSAGING$.notifySuccess('Suggestion successfully applied.');
-      setAppliedSuggestion(type);
-      setApplied([
-        ...applied,
-        {
-          [type]: selectedEntity[type],
-        },
-      ]);
-      setApplying(applying.filter((n) => n !== type));
-      if (onApplied) {
-        return onApplied(createdRelationships);
-      }
-    }
   };
 
   const settingsMessagesBannerHeight = useSettingsMessagesBannerHeight();
@@ -924,164 +651,14 @@ const ContainerHeader = (props) => {
               </Security>
             )}
             {enableSuggestions && (
-              <QueryRenderer
-                query={containerHeaderObjectsQuery}
-                variables={{ id: container.id }}
-                render={({ props: containerProps }) => {
-                  if (containerProps && containerProps.container) {
-                    const suggestions = generateSuggestions(
-                      containerProps.container.objects.edges.map((o) => ({
-                        ...o.node,
-                        types: o.types,
-                      })),
-                    );
-                    const appliedSuggestions = getAppliedSuggestions();
-                    if (userIsKnowledgeEditor) {
-                      return (
-                        <div style={{ marginLeft: theme.spacing(2) }}>
-                          <Tooltip title={t_i18n('Open the suggestions')}>
-                            <ToggleButton
-                              onClick={() => setDisplaySuggestions(true)}
-                              disabled={
-                                suggestions.length === 0
-                                || currentMode !== 'graph'
-                              }
-                              value="suggestion"
-                              size="small"
-                            >
-                              <Badge
-                                badgeContent={
-                                  suggestions.filter(
-                                    (n) => !appliedSuggestions.includes(n.type),
-                                  ).length
-                                }
-                                color="secondary"
-                              >
-                                <AssistantOutlined
-                                  fontSize="small"
-                                  disabled={suggestions.length === 0}
-                                  color={
-                                    // eslint-disable-next-line no-nested-ternary
-                                    suggestions.length === 0
-                                      ? 'disabled'
-                                      : displaySuggestions
-                                        ? 'secondary'
-                                        : 'primary'
-                                  }
-                                />
-                              </Badge>
-                            </ToggleButton>
-                          </Tooltip>
-                          <Dialog
-                            PaperProps={{ elevation: 1 }}
-                            open={displaySuggestions}
-                            TransitionComponent={Transition}
-                            onClose={() => setDisplaySuggestions(false)}
-                            maxWidth="md"
-                            fullWidth={true}
-                          >
-                            <DialogTitle>{t_i18n('Suggestions')}</DialogTitle>
-                            <DialogContent dividers={true}>
-                              <List>
-                                {suggestions.map((suggestion) => (
-                                  <ListItem
-                                    key={suggestion.type}
-                                    disableGutters={true}
-                                    divider={true}
-                                  >
-                                    <ListItemText
-                                      primary={
-                                        <MarkdownDisplay
-                                          content={t_i18n(
-                                            `suggestion_${suggestion.type}`,
-                                          )}
-                                          remarkGfmPlugin={true}
-                                          commonmark={true}
-                                          markdownComponents={true}
-                                        />
-                                      }
-                                    />
-                                    <Select
-                                      style={{
-                                        width: 200,
-                                        minWidth: 200,
-                                        margin: '0 0 0 15px',
-                                      }}
-                                      variant="standard"
-                                      onChange={(event) => handleSelectEntity(suggestion.type, event)
-                                      }
-                                      value={selectedEntity[suggestion.type]}
-                                    >
-                                      {suggestion.data.map((object) => (
-                                        <MenuItem
-                                          key={object.id}
-                                          value={object.id}
-                                        >
-                                          {getMainRepresentative(object)}
-                                        </MenuItem>
-                                      ))}
-                                    </Select>
-                                    <ListItemSecondaryAction>
-                                      <Tooltip title={t_i18n('Apply the suggestion')}>
-                                        <IconButton
-                                          edge="end"
-                                          aria-label="apply"
-                                          onClick={() => applySuggestion(
-                                            suggestion.type,
-                                            containerProps.container.objects.edges.map(
-                                              (o) => ({
-                                                ...o.node,
-                                                types: o.types,
-                                              }),
-                                            ),
-                                          )
-                                        }
-                                          size="large"
-                                          color={
-                                          applied.some(
-                                            (a) => a[suggestion.type]
-                                              === selectedEntity[suggestion.type],
-                                          )
-                                            ? 'success'
-                                            : 'primary'
-                                        }
-                                          disabled={
-                                          applying.includes(suggestion.type)
-                                          || !selectedEntity[suggestion.type]
-                                        }
-                                        >
-
-                                          {applying.includes(suggestion.type) ? (
-                                            <CircularProgress
-                                              size={20}
-                                              color="inherit"
-                                            />
-                                          ) : (
-                                            <AddTaskOutlined />
-                                          )}
-                                        </IconButton>
-                                      </Tooltip>
-                                    </ListItemSecondaryAction>
-                                  </ListItem>
-                                ))}
-                              </List>
-                            </DialogContent>
-                            <DialogActions>
-                              <Button
-                                onClick={() => setDisplaySuggestions(false)}
-                                color="primary"
-                              >
-                                {t_i18n('Close')}
-                              </Button>
-                            </DialogActions>
-                          </Dialog>
-                        </div>
-                      );
-                    }
-                  }
-                  return <div />;
-                }}
-              />
+            <StixCoreObjectsSuggestions
+              containerId={container.id}
+              currentMode={currentMode}
+              onApplied={onApplied}
+              instanceName={getMainRepresentative(container)}
+              containerHeaderObjectsQuery={containerHeaderObjectsQuery}
+              container={container}
+            />
             )}
             {enableQuickSubscription && (
               <StixCoreObjectQuickSubscription
